Extract line framing into a socket-free NntpLineFramer
NntpConnection had crossed 1k lines by mixing transport/session, line framing
and compression. Move the line-framing/data-block cluster (ReadLineAsync,
TryReadLine, DecodeLine, ReadDataBlockToBufferAsync, AppendLine(s),
EnsureCapacity, ProcessLine) into NntpLineFramer: pure byte->line logic over a
swappable PipeReader, with its own read-byte counter and read-fault flag.

The connection now holds a framer and swaps its Reader when rebuilding the
transport for compression. HasError is OR'd from the connection's write/response
faults and the framer's read faults. The framer is unit-tested over an in-memory
pipe with no socket. NntpConnection drops from ~1015 to ~745 lines.

Counter overflow now resets only the overflowing counter (read counting lives in
the framer, writes in the connection), which also addresses the surprising
both-counter reset noted in the review.
